refactor(tests): extract rename.rs (2 tests)
Move rename_* tests from mod.rs into engine/tests/rename.rs.

Tests moved (now at engine::tests::rename::*):
  rename_uses_named_realm
  rename_structured_doc_returns_not_supported_error

Tests: 135 passed / 0 failed / 1 ignored (baseline unchanged).

marky-n1h step 8 (2/5).
